test: mocking 제거 및 테스트 클래스 변경
